% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Analysis.R
\name{email}
\alias{email}
\title{Email results}
\usage{
email(from, to = getDestinationAddress(),
  subject = "OHDSI Skeleton Study Results", dataDescription,
  folder = getwd(), compress = TRUE)
}
\arguments{
\item{from}{Return email address}

\item{to}{Delivery email address (must be a gmail.com acccount)}

\item{subject}{Subject line of email}

\item{dataDescription}{A short description of the database}

\item{folder}{The name of the local folder to place results;  make sure to use forward slashes (/)}

\item{compress}{Use GZip compression on transmitted files}

\item{sourceName}{Short name that was be appeneded to results table name}
}
\value{
A list of files that were emailed.
}
\description{
Email results
}
\details{
This function emails the result CSV files to the study coordinator.
}

